server/requirements.txt
fastapi
uvicorn[standard]
httpx
chromadb
sentence-transformers
numpy
python-multipart
pandas
requests
python-dotenv
nltk
